requirements.txt
attrs==19.1.0
Django==1.11.17
pytz==2019.1
